+++ src/solve.rs
@@ -1,4 +1,4 @@
-use std::fmt::Display;
+use std::fmt::{Debug, Display};
 
 use anyhow::Result;
 
@@ -6,8 +6,9 @@ use crate::Input;
 
 /// A trait for solving an Advent of Code puzzle.
 ///
-/// This trait is automatically implemented for functions that take a reference to
-/// an [`Input`] and return a [`Result`] containing a type that implements [`Display`]:
+/// This trait is automatically implemented for functions that take a reference
+/// to an [`Input`] and return a [`Result`] containing a type that implements
+/// [`Debug`], [`Display`], and [`PartialEq`]:
 ///
 /// ```
 /// use adventurous::Input;
@@ -19,7 +20,7 @@ use crate::Input;
 /// ```
 pub trait Solve {
     /// The answer to the puzzle.
-    type Answer: Display;
+    type Answer: Debug + Display + PartialEq;
 
     /// Produces an answer from the provided [`Input`].
     fn solve(&self, input: &Input) -> Result<Self::Answer>;
@@ -27,7 +28,7 @@ pub trait Solve {
 
 impl<A, F> Solve for F
 where
-    A: Display,
+    A: Debug + Display + PartialEq,
     F: Fn(&Input) -> Result<A>,
 {
     type Answer = A;